Update device search page
